% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_retest.R
\name{plot_retest}
\alias{plot_retest}
\title{Plot temperatures}
\usage{
plot_retest(x, y)
}
\arguments{
\item{x}{x axis}

\item{y}{y axis}
}
\value{
Line plot
}
\description{
Plot temperatures
}
\examples{
plot(x=time, y=temp)


}
