% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedSearch.R
\name{pubmed_search}
\alias{pubmed_search}
\title{Pubmed Search}
\usage{
pubmed_search(keyword)
}
\arguments{
\item{keyword}{Character string of pubmed search}
}
\value{
Number of results from a search of pubmed database
}
\description{
Pubmed Search
}
\examples{
search("nursing[MeSH]")


}
