% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stargazer2.R
\name{stargazer2}
\alias{stargazer2}
\title{Logistic regression table with odd ratios}
\usage{
stargazer2(model, odd.ratio = TRUE, ...)
}
\arguments{
\item{model}{A non-empty list containing one or more glm models}

\item{odd.ratio}{A logical value indicating whether you want odd ratios (TRUE)
instead of the usual log odds (FALSE).}

\item{...}{further aguments passed to the stargazer function. See ?stargazer::stargazer}
}
\value{
A stargazer table with correct odd ratios, standard errors and p values
}
\description{
Logistic regression table with odd ratios
}
